Fixes a crash from ItemRenderer, fixes some issues with recipes
The addition to ContainerBlockWiki fixes some problems with rendering
items such as buttons into guis from auto-retrieved recipes.  It also
fixes issues with wildcard values, but a better way to handle this would
probably be needed later.

GuiWiki change is to fix a strange offset bug of some rendered tab items
which appeared in the 1.8 update.

The change to BlockAndItemWikiTab fixes a crash due to the Minecraft
RenderItem being null when that class is first referenced.

the change in IntegratorCraftingRecipe moves the text to the correct
spot.